+++ src/plugin/Properties/AssemblyInfo.cs
@@ -4,5 +4,5 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTitle("CyberArk.Extensions.Plugin.Python4CPM")]
 [assembly: Guid("0ebaf0ea-d0de-4bd7-8db7-8af58235114a")]
 [assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.1.4.0")]
-[assembly: AssemblyInformationalVersion("1.1.4")]
+[assembly: AssemblyFileVersion("1.1.5.0")]
+[assembly: AssemblyInformationalVersion("1.1.5")]