+++ src/python4cpm/logger.py
@@ -23,11 +23,11 @@ class Logger:
         os.makedirs(cls._LOGS_DIR, exist_ok=True)
         uid = os.urandom(4).hex()
         logger = logging.getLogger(uid)
-        _logging_level = logging_level.lower()
+        _logging_level = (logging_level or cls._DEFAULT_LEVEL).lower()
         if _logging_level not in cls._LOGGING_LEVELS:
             _logging_level = cls._DEFAULT_LEVEL
         logger.setLevel(cls._LOGGING_LEVELS[_logging_level])
-        file_name = f"{__name__}_{_logging_level}_{name}.log"
+        file_name = f"{__name__}_{_logging_level}_{name}_{uid}.log"
         logs_file = os.path.join(cls._LOGS_DIR, file_name)
         handler = RotatingFileHandler(
             filename=logs_file,
@@ -35,7 +35,7 @@ class Logger:
             backupCount=1
         )
         fmt = (
-            "%(asctime)s.%(msecs)03d | %(levelname)s | %(name)s | "
+            "%(asctime)s.%(msecs)03d | %(levelname)s | "
             "%(module)s | %(funcName)s | %(message)s"
         )
         datefmt = "%Y-%m-%d %H:%M:%S"